fix: Rename useForm hook to useProjectDetail for better clarity
